+++ src/guards/string/string-is-alpha.ts
@@ -6,13 +6,50 @@ import { GuardResult } from '../../core/guard-result';
 /**
  * Basic Latin pattern:
  * - Not case sensitive.
+ * @see {@link https://en.wikipedia.org/wiki/ISO_basic_Latin_alphabet}
  */
-export const BASIC_LATIN_PATTERN = '^[a-zA-Z]+$';
+const BASIC_LATIN_PATTERN = 'a-zA-Z';
+
+/**
+ * Latin Latin-1 Supplement pattern.
+ * @see {@link https://en.wikipedia.org/wiki/Latin-1_Supplement_(Unicode_block)}
+ */
+const LATIN_1_SUPPLEMENT_PATTERN = `À-ÖØ-öø-ÿ`;
+
+/**
+ * Latin Extended-A pattern.
+ * @see {@link https://en.wikipedia.org/wiki/Latin_Extended-A}
+ */
+const LATIN_EXTENDED_A_PATTERN = `Ā-ſ`;
+
+/**
+ * Latin Extended-B pattern.
+ * @see {@link https://en.wikipedia.org/wiki/Latin_Extended-B}
+ */
+const LATIN_EXTENDED_B_PATTERN = `ƀ-ɏ`;
+
+/**
+ * Latin Extended Additional pattern.
+ * @see {@link https://en.wikipedia.org/wiki/Latin_Extended_Additional}
+ */
+const LATIN_EXTENDED_ADDITIONAL_PATTERN = `Ḁ-ỿ`;
+
+/**
+ * @see {@link https://en.wikipedia.org/wiki/IPA_Extensions}
+ * @see {@link https://fr.wikipedia.org/wiki/Lettres_suppl%C3%A9mentaires_de_l%27alphabet_latins}
+ */
+const IPA_EXTENSIONS_PATTERN = `ʒ`;
+
+/**
+ * Latin trigram pattern.
+ * @see {@link https://fr.wikipedia.org/wiki/Liste_de_digrammes_et_trigrammes}
+ */
+const TRIGRAM_PATTERN = `((CʼH)|(Cʼh)|(cʼh))*`;
 
 /**
  * Partial precomposed Latin pattern.
  */
-const PARTIAL_PRECOMPOSED_LATIN_PATTERN = '^[a-zA-ZÀ-ÖØ-öø-ŸʒƷſß]+$';
+const PARTIAL_PRECOMPOSED_LATIN_PATTERN = `^[${BASIC_LATIN_PATTERN}${LATIN_1_SUPPLEMENT_PATTERN}${LATIN_EXTENDED_A_PATTERN}${LATIN_EXTENDED_B_PATTERN}${LATIN_EXTENDED_ADDITIONAL_PATTERN}${IPA_EXTENSIONS_PATTERN}]+$`;
 
 export class StringIsAlpha extends StringRuleChecker<{ type: 'isAlpha'; alphabet?: Alphabet }> {
     constructor(rule: { type: 'isAlpha'; alphabet?: Alphabet }, value: string) {
@@ -47,7 +84,7 @@ export class StringIsAlpha extends StringRuleChecker<{ type: 'isAlpha'; alphabet
     }
 
     private isBasicLatin(): boolean {
-        return this.value.match(new RegExp(BASIC_LATIN_PATTERN)) !== null;
+        return this.value.match(new RegExp(`^[${BASIC_LATIN_PATTERN}]+$`)) !== null;
     }
 
     private isPrecomposedLatin(): boolean {